Pickup Skill fully implemented
* Skill equip 1,2 correspondence implemented.
- Slots will be highlighted red/blue according to player's adjustment.
* Created function for skill damage calculation

+++ Assets/Scripts/Gameplay/Inventory/Equip.cs
@@ -7,5 +7,5 @@ public class Equip : Item
 {
     public string SkillAnimation;
     public List <GameObject> HitEffects;
-    public int skillDamage;
+    public float skillDamage;
 }
